+++ eslint-plugin/tests/lib/standalone.test.js
@@ -1,6 +1,6 @@
 /*
- * ecoCode JavaScript plugin - Provides rules to reduce the environmental footprint of your JavaScript programs
- * Copyright © 2023 Green Code Initiative (https://www.ecocode.io)
+ * creedengo JavaScript plugin - Provides rules to reduce the environmental footprint of your JavaScript programs
+ * Copyright © 2023 Green Code Initiative (https://green-code-initiative.org)
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -31,7 +31,7 @@ describe("standalone.js", () => {
     const recommended = configs.recommended;
     assert.notEqual(recommended, null);
     assert.equal(recommended.plugins.length, 1);
-    assert.equal(recommended.plugins[0], "@ecocode");
+    assert.equal(recommended.plugins[0], "@creedengo");
     assert.equal(recommended.rules.length, rules.length);
   });
 });